requirements.txt
swig
faiss-cpu
langchain
PyPDF2
python-dotenv
streamlit
streamlit-ace
tiktoken
openai
beautifulsoup4
